# swagger.yaml
openapi: 3.0.0
info:
  title: SOA Test BE
  description: APIs for SOA Test BE
  version: 1.0.0
servers:
  - url: https://{host}/{basePath}
    description: Production server
    variables:
      host:
        default: api.example.com
        description: API host
      basePath:
        default: api/soa/v1
        description: Base path for all API endpoints
paths:
...
  /categories:
    post:
      summary: Create a new category
      description: Creates a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '200':
          description: category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    properties:
                      record:
                        $ref: '#/components/schemas/Category'
                    required:
                      - record
                required:
                  - meta
                  - data
        '422':
          description: Invalid input
...
  /cities:
    post:
      summary: Create a new city
      description: Creates a new city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '200':
          description: city created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    properties:
                      record:
                        $ref: '#/components/schemas/City'
                    required:
                      - record
                required:
                  - meta
                  - data
        '422':
          description: Invalid input
...
  /suppliers:
    post:
      summary: Create a new supplier
      description: Creates a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '200':
          description: suppliers created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    properties:
                      record:
                        $ref: '#/components/schemas/Supplier'
                    required:
                      - record
                required:
                  - meta
                  - data
        '422':
          description: Invalid input
...
  /products:
    post:
      summary: Create a new product
      description: Creates a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '200':
          description: product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    properties:
                      record:
                        $ref: '#/components/schemas/Product'
                    required:
                      - record
                required:
                  - meta
                  - data
        '422':
          description: Invalid input
...
components:
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
          example: "1aaf47cc-b4ea-4a4c-9fb7-0768cc7de128"
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        time:
          type: string
          format: date-time
          example: "2025-03-16 15:35:37"
        pagination:
          type: object
          properties:
            limit:
              type: integer
              default: 20
            offset:
              type: integer
              default: 0
      required:
        - request_id
        - code
        - message
        - time
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "5268d217-58f6-4d83-a7a1-a52605e3786a"
        name:
          type: string
          example: "Electronics"
        status:
          type: integer
          example: 1
        created_time:
          type: string
          format: date-time
          example: "2025-03-15T13:07:28.437235Z"
        updated_time:
          type: string
          format: date-time
          example: "2025-03-15T13:07:28.437235Z"
        deleted_time:
          type: string
          format: date-time
          nullable: true
          example: null
      required:
        - id
        - name
        - status
        - created_time
        - updated_time
    CategoryCreate:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
    City:
      type: object
      properties:
        ID:
          type: string
          format: uuid
          example: "dd6bec56-e7cf-4ef8-8488-0e41281c91e8"
          description: Unique identifier for the location
        name:
          type: string
          example: "Maple Grove"
          description: Name of the location
        latitude:
          type: number
          format: float
          example: 45.0724
          description: Latitude coordinate of the location
        longitude:
          type: number
          format: float
          example: 93.4557
          description: Longitude coordinate of the location
        created_time:
          type: string
          format: date-time
          example: "2025-03-15T13:20:22.531526Z"
          description: Timestamp when the location was created
        updated_time:
          type: string
          format: date-time
          example: "2025-03-15T13:20:22.531526Z"
          description: Timestamp when the location was last updated
        deleted_time:
          type: string
          format: date-time
          nullable: true
          example: null
          description: Timestamp when the location was deleted, null if not deleted
      required:
        - ID
        - name
        - latitude
        - longitude
        - created_time
        - updated_time
...
    Supplier:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "5268d217-58f6-4d83-a7a1-a52605e3786a"
        name:
          type: string
          example: "Demo"
        status:
          type: integer
          example: 1
        created_time:
          type: string
          format: date-time
          example: "2025-03-15T13:07:28.437235Z"
        updated_time:
          type: string
          format: date-time
          example: "2025-03-15T13:07:28.437235Z"
        deleted_time:
          type: string
          format: date-time
          nullable: true
          example: null
      required:
        - id
        - name
        - status
        - created_time
        - updated_time
    SupplierCreate:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c5ec2b98-b9c6-41d3-9514-2c7c4575c4a1"
          description: Unique identifier for the product
        reference:
          type: string
          example: "PROD-202503-001"
          description: Product reference code
        name:
          type: string
          example: "demo"
          description: Name of the product
        status:
          type: integer
          example: 1
          description: Status of the product (e.g., 1 for active)
        price:
          type: string
          example: "300"
          description: Price of the product
        quantity:
          type: integer
          example: 20
          description: Available quantity of the product
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
          description: Categories associated with the product
        city_id:
          type: string
          format: uuid
          example: "dd6bec56-e7cf-4ef8-8488-0e41281c91e8"
          description: ID of the city where the product is located
        city:
          $ref: '#/components/schemas/City'
        supplier_id:
          type: string
          format: uuid
          example: "324e390b-d895-40b6-9f61-fb3339aa3a47"
          description: ID of the product supplier
        supplier:
          $ref: '#/components/schemas/Supplier'
        created_time:
          type: string
          format: date-time
          example: "2025-03-16T10:11:46.988146Z"
          description: Timestamp when the product was created
        updated_time:
          type: string
          format: date-time
          example: "2025-03-16T10:11:46.988146Z"
          description: Timestamp when the product was last updated
        deleted_time:
          type: string
          format: date-time
          nullable: true
          example: null
          description: Timestamp when the product was deleted, null if not deleted
      required:
        - id
        - name
        - status
        - created_time
        - updated_time
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          example: "demo"
        price:
          type: number
          format: float
          example: 300
        quantity:
          type: integer
          example: 20
        status:
          type: integer
          example: 1
        categories:
          type: array
          items:
            type: string
          example:
            - "5268d217-58f6-4d83-a7a1-a52605e3786a"
            - "23bae8b1-1c4a-4cf1-a8b9-193fa98a094b"
        city_id:
          type: string
          format: uuid
          example: "dd6bec56-e7cf-4ef8-8488-0e41281c91e8"
        supplier_id:
          type: string
          format: uuid
          example: "324e390b-d895-40b6-9f61-fb3339aa3a47"
      required:
        - name
        - price
        - quantity
        - status
        - categories
        - city_id
        - supplier_id
...